+++ packages/stream_chat/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - You can now pass `score` to `client.sendReaction` and `channel.sendReaction` functions
 
+🐞 Fixed
+
+- [[#890]](https://github.com/GetStream/stream-chat-flutter/pull/890). Fixed Reactions not updating on thread messages. Thanks [bstolinski](https://github.com/bstolinski).
+- 
 ## 3.4.0
 
 🐞 Fixed